#3 реализовал операции редактирования и добавления. Вынес в отдельный класс оперции с бд.
